Get basic router working
