+++ timestrap/static_src/tags/clients.vue
@@ -1,0 +1,10 @@
+<template>
+<div class="content">
+    <div class="container">
+        Hello!
+    </div>
+</div>
+</template>
+
+<script>
+</script>
