+++ timestrap/static_src/tags/main.js
@@ -1,12 +1,32 @@
-var Vue = require('vue')
-var App = require('./app.vue')
+const Vue = require('vue');
+const VueRouter = require('vue-router');
 
-new Vue({
+const App = require('./app.vue');
+const Timesheet = require('./timesheet.vue');
+const Clients = require('./clients.vue');
+
+
+Vue.use(VueRouter);
+
+
+const routes = [
+    { path: '/timesheet/', component: Timesheet },
+    { path: '/clients/', component: Clients }
+];
+
+
+const router = new VueRouter({
+    mode: 'history',
+    hasbang: false,
+    linkActiveClass: 'active',
+    routes
+});
+
+
+const app = new Vue({
+    router,
     el: '#app',
-    render: function (createElement) {
-        return createElement(App)
-    },
-    use: {
-        VueRouter
+    render(createElement) {
+        return createElement(App);
     }
 });
